requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.123.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
ffmpy==1.0.0
filelock==3.20.0
fsspec==2025.10.0
google-auth==2.43.0
google-genai==1.52.0
gradio==6.0.2
gradio_client==2.0.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.7
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.5
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
safehttpx==0.1.7
semantic-version==2.10.0
sentry-sdk==2.46.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
